type specimen project
I tried to explore variety of things through this project. Not just for
the required assignment including justification and change font styles,
I figured out how to deal with the table and change the color or line
of it. Also I learned how to do some drop crop of the first letter.